Move algebraic structures into their own folder
