+++ packages/core/package.json
@@ -35,14 +35,11 @@
     "modules": [
       "Async",
       "Array",
-      "Associative",
       "Boolean",
       "Bounded",
       "Branded",
       "Case",
       "Chunk",
-      "Closure",
-      "Commutative",
       "Const",
       "Dictionary",
       "Either",
@@ -51,10 +48,8 @@
       "FreeAssociative",
       "Hash",
       "Id",
-      "Identity",
       "IndexedT",
       "InvariantT",
-      "Inverse",
       "IO",
       "Iterable",
       "Map",
@@ -154,6 +149,11 @@
       "Prelude/WiltableWithIndex",
       "Prelude/Witherable",
       "Prelude/WitherableWithIndex",
+      "Structure/Associative",
+      "Structure/Closure",
+      "Structure/Commutative",
+      "Structure/Identity",
+      "Structure/Inverse",
       "Sync",
       "Support",
       "Support/AtomicBoolean",
