+++ packages/core/src/Array/operations.ts
@@ -9,14 +9,14 @@ import type { MutableArray } from "@effect-ts/system/Mutable"
 import type { Either } from "../Either"
 import type { Equal } from "../Equal"
 import { makeEqual } from "../Equal"
-import type { Identity } from "../Identity"
-import { makeIdentity } from "../Identity"
 import type { ArrayURI } from "../Modules"
 import * as Ord from "../Ord"
 import type { URI } from "../Prelude"
 import * as P from "../Prelude"
 import * as DSL from "../Prelude/DSL"
 import type { Show } from "../Show"
+import type { Identity } from "../Structure/Identity"
+import { makeIdentity } from "../Structure/Identity"
 import type { PredicateWithIndex, Separated } from "../Utils"
 
 export * from "@effect-ts/system/Array"
