+++ packages/core/src/Dictionary/operations.ts
@@ -5,20 +5,20 @@ import type { MutableRecord } from "@effect-ts/system/Mutable"
 import * as O from "@effect-ts/system/Option"
 
 import * as A from "../Array"
-import type { Associative } from "../Associative"
-import type { Closure } from "../Closure"
 import * as E from "../Either"
 import type { Equal } from "../Equal"
 import { makeEqual } from "../Equal"
 import { identity, pipe, tuple } from "../Function"
-import type { Identity } from "../Identity"
-import { makeIdentity } from "../Identity"
 import type { DictionaryURI } from "../Modules"
 import type { Foldable, URI } from "../Prelude"
 import * as P from "../Prelude"
 import { succeedF } from "../Prelude"
 import type * as HKT from "../Prelude/HKT"
 import type { Show } from "../Show"
+import type { Associative } from "../Structure/Associative"
+import type { Closure } from "../Structure/Closure"
+import type { Identity } from "../Structure/Identity"
+import { makeIdentity } from "../Structure/Identity"
 
 export * from "@effect-ts/system/Dictionary"
