+++ packages/core/src/Effect/index.ts
@@ -4,11 +4,11 @@ import "../Operator"
 
 import * as T from "@effect-ts/system/Effect"
 
-import * as I from "../Identity"
 import type { EffectCategoryURI, EffectURI } from "../Modules"
 import type { URI } from "../Prelude"
 import * as P from "../Prelude"
 import * as DSL from "../Prelude/DSL"
+import * as I from "../Structure/Identity"
 
 export * from "@effect-ts/system/Effect"
 export { EffectURI } from "../Modules"
