+++ packages/core/src/Either/operations.ts
@@ -4,16 +4,16 @@ import type { Either } from "@effect-ts/system/Either"
 import * as E from "@effect-ts/system/Either"
 import { pipe, tuple } from "@effect-ts/system/Function"
 
-import type { Associative } from "../Associative"
-import { makeAssociative } from "../Associative"
 import type { Equal } from "../Equal"
-import type { Identity } from "../Identity"
 import type { EitherURI } from "../Modules"
 import type { Option } from "../Option"
 import type { URI } from "../Prelude"
 import * as P from "../Prelude"
 import * as DSL from "../Prelude/DSL"
 import type { Show } from "../Show"
+import type { Associative } from "../Structure/Associative"
+import { makeAssociative } from "../Structure/Associative"
+import type { Identity } from "../Structure/Identity"
 import type { Separated } from "../Utils"
 import type { V } from "./definition"
