+++ packages/core/src/FreeAssociative/instances.ts
@@ -2,11 +2,11 @@
 
 import * as FA from "@effect-ts/system/FreeAssociative"
 
-import { makeAssociative } from "../Associative"
-import { makeIdentity } from "../Identity"
 import type { FreeAssociativeURI } from "../Modules"
 import type { URI } from "../Prelude"
 import * as P from "../Prelude"
+import { makeAssociative } from "../Structure/Associative"
+import { makeIdentity } from "../Structure/Identity"
 
 export function getAssociative<A>() {
   return makeAssociative<FA.FreeAssociative<A>>(FA.concat_)
