+++ packages/core/src/Id/index.ts
@@ -4,12 +4,12 @@
 import "../Operator"
 
 import type { Equal } from "../Equal"
-import type { Identity } from "../Identity"
 import type { IdURI } from "../Modules"
 import type { URI } from "../Prelude"
 import * as P from "../Prelude"
 import { structF, tupleF } from "../Prelude/DSL"
 import type { Show } from "../Show"
+import type { Identity } from "../Structure/Identity"
 
 export type Id<A> = A
