+++ packages/core/src/List/operations.ts
@@ -6,13 +6,13 @@ import * as A from "../Array"
 import type { Equal } from "../Equal"
 import { makeEqual } from "../Equal"
 import { pipe } from "../Function"
-import type { Identity } from "../Identity"
-import { makeIdentity } from "../Identity"
 import type { ListURI } from "../Modules"
 import type { Ord } from "../Ord"
 import type { URI } from "../Prelude"
 import * as P from "../Prelude"
 import type { Show } from "../Show"
+import type { Identity } from "../Structure/Identity"
+import { makeIdentity } from "../Structure/Identity"
 
 export * from "@effect-ts/system/Persistent/List"
