+++ packages/core/src/Modules/index.ts
@@ -7,8 +7,6 @@ import type { Effect, EffectURI } from "@effect-ts/system/Effect"
 import type { Array } from "../Array"
 import type { Async } from "../Async"
 import type { Bounded } from "../Bounded"
-import type { Closure } from "../Closure"
-import type { Commutative } from "../Commutative"
 import type { Const } from "../Const"
 import type { Dictionary } from "../Dictionary"
 import type { Layer } from "../Effect/Layer"
@@ -16,9 +14,7 @@ import type { Either } from "../Either"
 import type { Equal } from "../Equal"
 import type { FreeAssociative } from "../FreeAssociative"
 import type { Id } from "../Id"
-import type { Identity } from "../Identity"
 import type { Ix, IxC } from "../IndexedT"
-import type { Inverse } from "../Inverse"
 import type { IO } from "../IO"
 import type { List } from "../List"
 import type { NonEmptyArray } from "../NonEmptyArray"
@@ -28,6 +24,10 @@ import type { Reader } from "../Reader"
 import type { Tree } from "../RoseTree"
 import type { Show } from "../Show"
 import type { StateIn, StateOut } from "../StateT/Classic"
+import type { Closure } from "../Structure/Closure"
+import type { Commutative } from "../Structure/Commutative"
+import type { Identity } from "../Structure/Identity"
+import type { Inverse } from "../Structure/Inverse"
 import type { Sync } from "../Sync"
 import type { XPure } from "../XPure"
 import type { XIO } from "../XPure/XIO"
