+++ packages/core/src/NonEmptyArray/operations.ts
@@ -7,16 +7,16 @@ import type { NonEmptyArray } from "@effect-ts/system/NonEmptyArray"
 import * as NA from "@effect-ts/system/NonEmptyArray"
 import * as L from "@effect-ts/system/Persistent/List"
 
-import { makeAssociative } from "../Associative"
 import type { Equal } from "../Equal"
 import { makeEqual } from "../Equal"
-import type { Identity } from "../Identity"
 import type { NonEmptyArrayURI } from "../Modules"
 import * as Ord from "../Ord"
 import type { URI } from "../Prelude"
 import * as P from "../Prelude"
 import * as DSL from "../Prelude/DSL"
 import type { Show } from "../Show"
+import { makeAssociative } from "../Structure/Associative"
+import type { Identity } from "../Structure/Identity"
 
 export * from "@effect-ts/system/NonEmptyArray"
