+++ packages/core/src/Option/operations.ts
@@ -2,21 +2,21 @@
 
 import * as O from "@effect-ts/system/Option"
 
-import type { Associative } from "../Associative"
-import { makeAssociative } from "../Associative"
 import type { Either } from "../Either"
 import { left, right } from "../Either"
 import type { Equal } from "../Equal"
 import type { Predicate } from "../Function"
 import { pipe } from "../Function"
-import type { Identity } from "../Identity"
-import { fold, fromAssociative, makeIdentity } from "../Identity"
 import type { OptionURI } from "../Modules"
 import type { Ord } from "../Ord"
 import { makeOrd } from "../Ord"
 import type { URI } from "../Prelude"
 import * as P from "../Prelude"
 import type { Show } from "../Show"
+import type { Associative } from "../Structure/Associative"
+import { makeAssociative } from "../Structure/Associative"
+import type { Identity } from "../Structure/Identity"
+import { fold, fromAssociative, makeIdentity } from "../Structure/Identity"
 import type { Separated } from "../Utils"
 
 export function getEqual<A>(E: Equal<A>): Equal<O.Option<A>> {
