+++ packages/core/src/Ordering/operations.ts
@@ -2,8 +2,8 @@
 
 import type { Ordering } from "@effect-ts/system/Ordering"
 
-import * as A from "../Associative/makeAssociative"
-import * as I from "../Identity/makeIdentity"
+import * as A from "../Structure/Associative/makeAssociative"
+import * as I from "../Structure/Identity/makeIdentity"
 
 /**
  * `Associative` instance for `Ordering`
