+++ packages/core/src/Prelude/FoldMap/index.ts
@@ -1,6 +1,6 @@
 // tracing: off
 
-import type { Identity } from "../../Identity"
+import type { Identity } from "../../Structure/Identity"
 import type * as HKT from "../HKT"
 
 export interface FoldMap<F extends HKT.URIS, C = HKT.Auto> extends HKT.Base<F, C> {
