+++ packages/core/src/Set/index.ts
@@ -6,17 +6,17 @@ import type { MutableSet } from "@effect-ts/system/Mutable"
 import type { Option } from "@effect-ts/system/Option"
 import type { Separated } from "@effect-ts/system/Utils"
 
-import type { Associative } from "../Associative"
-import { makeAssociative } from "../Associative"
 import type { Either } from "../Either"
 import type { Equal } from "../Equal"
 import { makeEqual } from "../Equal"
 import type { Predicate, Refinement } from "../Function"
 import { identity, not } from "../Function"
-import type { Identity } from "../Identity"
-import { makeIdentity } from "../Identity"
 import type { Ord } from "../Ord"
 import type { Show } from "../Show"
+import type { Associative } from "../Structure/Associative"
+import { makeAssociative } from "../Structure/Associative"
+import type { Identity } from "../Structure/Identity"
+import { makeIdentity } from "../Structure/Identity"
 
 export type Set<A> = ReadonlySet<A>
