+++ packages/core/src/Structure/Associative/common.ts
