+++ packages/core/src/Structure/Associative/definition.ts
