+++ packages/core/src/Structure/Associative/makeAssociative.ts
@@ -1,6 +1,6 @@
 // tracing: off
 
-import { instance } from "../Prelude"
+import { instance } from "../../Prelude"
 import type { Associative } from "./definition"
 
 export function makeAssociative<A>(f: (x: A, y: A) => A): Associative<A> {
