+++ packages/core/src/Structure/Closure/operations.ts
@@ -1,6 +1,6 @@
 // tracing: off
 
-import { instance } from "../Prelude"
+import { instance } from "../../Prelude"
 import type { Closure } from "./definition"
 
 export function makeClosure<A>(f: (x: A, y: A) => A): Closure<A> {
