+++ packages/core/src/Structure/Commutative/index.ts
@@ -1,7 +1,7 @@
 // tracing: off
 
-import "../Operator"
+import "../../Operator"
 
 export * from "./definition"
-export { InverseURI } from "../Modules"
+export { CommutativeURI } from "../../Modules"
 export * from "./operations"
