+++ packages/core/src/Structure/Commutative/operations.ts
@@ -1,6 +1,6 @@
 // tracing: off
 
-import { instance } from "../Prelude"
+import { instance } from "../../Prelude"
 import type { Commutative } from "./definition"
 
 export function makeCommutative<A>(f: (l: A, r: A) => A): Commutative<A> {
