+++ packages/core/src/Structure/Identity/common.ts
