+++ packages/core/src/Structure/Inverse/definition.ts
