+++ packages/core/src/index.ts
@@ -3,13 +3,13 @@
 import "./Operator"
 
 export * as Array from "./Array"
-export * as Associative from "./Associative"
+export * as Associative from "./Structure/Associative"
 export * as Async from "./Async"
 export * as Boolean from "./Boolean"
 export * as Branded from "./Branded"
 export * as Chunk from "./Chunk"
-export * as Closure from "./Closure"
-export * as Commutative from "./Commutative"
+export * as Closure from "./Structure/Closure"
+export * as Commutative from "./Structure/Commutative"
 export * as Const from "./Const"
 export * as Effect from "./Effect"
 export * as Cause from "./Effect/Cause"
@@ -46,10 +46,10 @@ export * as HashMap from "./HashMap"
 export * as HashSet from "./HashSet"
 export * as IO from "./IO"
 export * as Id from "./Id"
-export * as Identity from "./Identity"
+export * as Identity from "./Structure/Identity"
 export * as IndexedT from "./IndexedT"
 export * as InvariantT from "./InvariantT"
-export * as Inverse from "./Inverse"
+export * as Inverse from "./Structure/Inverse"
 export * as Iterable from "./Iterable"
 export * as List from "./List"
 export * as Map from "./Map"
