+++ packages/core/test/examples/08.test.ts
@@ -1,8 +1,8 @@
 import { pipe } from "@effect-ts/system/Function"
 
-import { makeAssociative } from "../../src/Associative"
 import * as T from "../../src/Effect"
 import * as DSL from "../../src/Prelude/DSL"
+import { makeAssociative } from "../../src/Structure/Associative"
 
 test("08", async () => {
   const ValidationApplicative = T.getValidationApplicative(
