+++ packages/core/test/examples/13.test.ts
@@ -6,13 +6,13 @@ import { pipe } from "@effect-ts/system/Function"
 import { NoSuchElementException } from "@effect-ts/system/GlobalExceptions"
 import { isEither, isOption } from "@effect-ts/system/Utils"
 
-import { makeAssociative } from "../../src/Associative"
 import * as Either from "../../src/Either"
 import * as EitherT from "../../src/EitherT"
 import * as Option from "../../src/Option"
 import { GenHKT } from "../../src/Prelude/DSL"
 import * as DSL from "../../src/Prelude/DSL"
 import * as S from "../../src/Prelude/Selective"
+import { makeAssociative } from "../../src/Structure/Associative"
 import type { XIO } from "../../src/XPure/XIO"
 import * as IO from "../../src/XPure/XIO"
 import * as Reader from "../../src/XPure/XReader"
